==> 3_preprocessing/1_data_loading.ipynb <==
# Step 1: Data Loading

## Preprocessing Pipeline Overview

This preprocessing pipeline outlines the steps necessary to prepare the Telco Customer Churn dataset for our modeling. Each step is designed to address specific aspects of data quality, transformation, and feature creation. We will cover each step in a separate jupyter notebook file.

**Step 1: Data Loading**: Loading the datasets into the workspace, ensuring all necessary files are correctly imported for analysis. This includes the Kaggle dataset and the IBM datasets.

**Step 2: Dataset Integration**: Combining relevant datasets into a single, unified dataset that will serve as the foundation for subsequent analysis.

**Step 3: Handling Missing Values**: Identifying and addressing missing values in the dataset to ensure data integrity. This step ensures no significant gaps hinder the analysis.

**Step 4: Data Type Conversion**: Converting data columns to appropriate data types to optimize memory usage and prepare for feature engineering. Ensure consistency across all columns.

**Step 5: Data Exploration**: Perform initial exploratory data analysis (EDA) to understand the dataset's structure and characteristics, visualizing key features to gain insights into the data.

**Step 6: Feature Engineering**: Creating new features from the existing data to enhance model performance and capture additional insights. This includes transformations and derived features.

**Step 7: Dataset Splitting**: Splitting the dataset into training and testing subsets to prepare for model development and evaluation. This step ensures reproducibility and robust performance metrics.

**Step 8: Outlier Detection**: Identifying and addressing outliers in the dataset to ensure they do not negatively impact the analysis or models.

**Step 9: Clustering Customers**: Identifying the most common customer profiles via clustering.cs.

import pandas as pd
import os

## Data Sources

In this analysis, we utilize two primary datasets:

1. **Kaggle Dataset**: This dataset provides customer information from a telecommunications company, including demographics, account details, and service usage.

2. **IBM Dataset**: This dataset offers more comprehensive and detailed data, including additional aspects such as demographics, location, population, services, and customer status. Furthermore, an already combined version exists (`telco_customer_churn.xlsx`).

By integrating these datasets, we get more detailed customer information for our analysis in the next steps of the preprocessing pipeline.

# Layout of the data
pd.set_option('display.max_columns', None)  # Show all columns
pd.set_option('display.expand_frame_repr', False)  # Prevent line wrapping

## 1.1 Original Kaggle Dataset

### General Information on the Kaggle Dataset

The `df_original` dataframe contains customer information from a telecommunications company. It includes 21 columns with details such as customer ID, gender, senior citizen status, partner status, dependents, tenure, phone service, multiple lines, internet service, online security, online backup, device protection, tech support, streaming TV, streaming movies, contract type, paperless billing, payment method, monthly charges, total charges, and churn status. This dataset is sourced from [Kaggle](https://www.kaggle.com/datasets/blastchar/telco-customer-churn).

# Load the dataset
file_path = '../2_data/original_data/Customer-Churn.csv'
df_original = pd.read_csv(file_path)

df_original.head()

# Display a concise summary of the DataFrame
df_original.info()

# Get statistical summaries of the numerical columns
df_original.describe()

### Analysis of the Original Kaggle Dataset

#### Key Insights:

1. **Customer Demographics**:
    - **Gender**: The dataset includes both male and female customers.
    - **Senior Citizens**: Approximately 16.2% of the customers are senior citizens.

2. **Account Information**:
    - **Partner and Dependents**: A significant portion of customers have partners, but fewer have dependents.
    - **Contract Types**: Customers have various contract types, including month-to-month, one-year, and two-year contracts.
    - **Payment Methods**: Customers use different payment methods, such as electronic check, mailed check, bank transfer, and credit card.

3. **Service Usage**:
    - **Phone and Internet Services**: The dataset includes customers with and without phone service, multiple lines, and different types of internet services (DSL, Fiber optic).
    - **Additional Services**: Customers subscribe to various additional services like online security, online backup, device protection, tech support, streaming TV, and streaming movies.

4. **Charges**:
    - **Monthly Charges**: The average monthly charge is approximately $64.76, with a wide range from $18.25 to $118.75.
    - **Total Charges**: The total charges vary significantly among customers, reflecting different service usage and tenure.

5. **Customer Tenure**:
    - **Tenure**: The tenure of customers ranges from 0 to 72 months, with a median of 29 months. This indicates a diverse range of customer loyalty.

6. **Churn**:
    - **Churn Rate**: The dataset includes information on whether customers have churned, which is crucial for predictive modeling and understanding customer retention.

### Summary Statistics:

- **SeniorCitizen**:
  - Mean: 0.162
  - Standard Deviation: 0.369
  - Min: 0
  - Max: 1

- **Tenure**:
  - Mean: 32.37 months
  - Standard Deviation: 24.56 months
  - Min: 0 months
  - Max: 72 months

- **MonthlyCharges**:
  - Mean: $64.76
  - Standard Deviation: $30.09
  - Min: $18.25
  - Max: $118.75

These insights provide a comprehensive overview of the customer demographics, account information, service usage, charges, tenure, and churn, which are essential for further analysis and modeling.

### Conclusion on the Original Kaggle Dataset and next steps

There is a newer version of the dataset available on the IBM website, which provides more comprehensive and detailed data. We will focus on this dataset for our analysis and modeling.

## 1.2 IBM Dataset

### General Information on the IBM Dataset

For more comprehensive data, we found additional datasets in a newer version available on the corresponding [IBM website](https://community.ibm.com/community/user/businessanalytics/blogs/steven-macko/2019/07/11/telco-customer-churn-1113). These datasets provide detailed information on various aspects such as demographics, location, population, services, and customer status. By integrating these datasets, we can enrich our analysis with more granular and comprehensive customer information.

The `(../2_data/Telecommunications_Industry)` folder contains various Excel files sourced from the IBM Website. Each file within this folder pertains to a specific topic relevant to the telecommunications industry. Below is a brief description of each topic covered by these files:

**Telco_customer_churn_demographics.xlsx**

This file contains demographic information about the customers, such as age, gender, and marital status. This data is useful for understanding the customer base and performing market segmentation.


**Telco_customer_churn_location.xlsx**

This file includes data on the geographical location of customers, including their state, city, and zip code. It helps in analyzing the distribution of customers across different regions.

**Telco_customer_churn_population.xlsx**

This file provides details on the population statistics related to the customers, such as population density and other relevant metrics. It helps in understanding the broader demographic context of the customer base.

**Telco_customer_churn_services.xlsx**

This file includes data on the various services subscribed to by customers, such as phone service, internet service, and additional features like online security and tech support. It helps in analyzing service usage patterns and customer preferences.

**Telco_customer_churn_status.xlsx**

This file captures the status of customers, including their tenure, contract type, payment method, and churn status. It is essential for understanding customer retention and identifying factors contributing to churn.

Each of these files provides valuable insights into different aspects of the telecommunications industry, enabling data-driven decision-making and strategic planning.


## 1.2.1 IBM Demographics Dataset

# Load the demographics dataset
demographics_file_path = '../2_data/original_data/Telecommunications_Industry/Telco_customer_churn_demographics.xlsx'
csv_file_path = '../2_data/Telco_customer_churn_demographics.csv'
df_demographics = pd.read_excel(demographics_file_path)

# Save the demographics dataset to a CSV file
df_demographics.to_csv(csv_file_path)

# Load the demographics dataset from the CSV file
df_demographics = pd.read_csv(csv_file_path)

# Display the first few rows of the demographics dataset
df_demographics.head()

### General Information on the IBM Demographics Dataset

The `telco_customer_churn_demographics.xlsx` dataset, abbreviated with `df_demographics` in our further analysis, provides detailed demographic information about the customers of a telecommunications company. This dataset includes the following key features.

# Display a concise summary of the demographics DataFrame
df_demographics.info()

# Get statistical summaries of the numerical columns in the demographics DataFrame
df_demographics.describe()

### Key Insights from Demographics Data

Our goal is to get a general understanding of this dataset. We will analyze it in detail later.

1. **Gender Distribution**:
    - The dataset includes both male and female customers, providing a balanced view of the customer base.

2. **Age Distribution**:
    - The average age of customers is approximately 46.5 years.
    - The age range spans from 19 to 80 years, indicating a diverse age group among the customers.
    - The standard deviation of 16.75 years suggests a wide variation in customer ages.

3. **Senior Citizens**:
    - The dataset includes information on whether customers are senior citizens (65 years or older). This can help in understanding the needs and preferences of older customers.

4. **Marital Status**:
    - The dataset captures whether customers are married, which can be useful for segmenting customers based on their family status.

5. **Dependents**:
    - Information on whether customers have dependents and the number of dependents they have is available.
    - The average number of dependents is approximately 0.47, with a maximum of 9 dependents.
    - A significant portion of customers do not have any dependents, as indicated by the 25th, 50th, and 75th percentiles all being 0.

6. **Customer Count**:
    - The `Count` column is used for reporting and dashboarding purposes, indicating the number of customers in a filtered set.

These insights provide a comprehensive overview of the customer demographics, which can be crucial for targeted marketing, customer segmentation, and personalized service offerings.

## 1.2.2 IBM Location Dataset

### General Information on the IBM Location Dataset

The `telco_customer_churn_location.xlsx` dataset, abbreviated with `df_location` in our further analysis, provides detailed location information about the customers of a telecommunications company. This data helps us understand the geographical distribution of customers, which can be crucial for regional marketing strategies, service optimization, and identifying location-specific customer needs. By analyzing the location data, we can gain insights into customer concentration in different areas, regional preferences, and potential areas for expanding services or improving customer support.

# Load the location dataset
location_file_path = '../2_data/original_data/Telecommunications_Industry/Telco_customer_churn_location.xlsx'
csv_file_path = '../2_data/Telco_customer_churn_location.csv'
df_location = pd.read_excel(location_file_path)

# Save the location dataset to a CSV file
df_location.to_csv(csv_file_path)

# Load the location dataset from the CSV file
df_location = pd.read_csv(csv_file_path)

# Display the first few rows of the location dataset
df_location.head()

# Display a concise summary of the demographics DataFrame
df_location.info()

# Get statistical summaries of the numerical columns in the demographics DataFrame
df_location.describe()

### Key Insights from Location Data

1. **Geographical Distribution**:
    - The dataset covers a wide range of zip codes, indicating a diverse geographical distribution of customers.
    - The latitude and longitude values suggest that customers are spread across various regions, with significant variation in their locations.

2. **Customer Concentration**:
    - The mean and median zip codes are close to each other, suggesting a relatively even distribution of customers across different areas.
    - The standard deviation of the zip codes indicates some variability, but not extreme, in the customer locations.

3. **State and City Representation**:
    - The dataset includes customers from multiple states and cities, providing a comprehensive view of the customer base across different regions.
    - This diversity can be useful for regional analysis and understanding location-specific trends.

4. **Latitude and Longitude Range**:
    - The range of latitude and longitude values shows that customers are located from the southernmost to the northernmost parts of the region, as well as from the easternmost to the westernmost parts.
    - This wide range can help in analyzing regional differences in customer behavior and preferences.

5. **Data Completeness**:
    - All columns have complete data with no missing values, ensuring the reliability of the location information for further analysis.

These insights highlight the geographical diversity and distribution of the customer base, which can be crucial for regional marketing strategies, service optimization, and understanding location-specific customer needs.

## 1.2.3 IBM Population Dataset

### General Information on the IBM Population Dataset

The `telco_customer_churn_population.xlsx` dataset, abbreviated with `df_population` in our further analysis, provides detailed population information related to the customers of a telecommunications company. This dataset helps in understanding the broader demographic context of the customer base, which can be useful for regional analysis and identifying trends based on population density.

# Load the population dataset
population_file_path = '../2_data/original_data/Telecommunications_Industry/Telco_customer_churn_population.xlsx'
csv_file_path = '../2_data/Telco_customer_churn_population.csv'
df_population = pd.read_excel(population_file_path)

# Save the population dataset to a CSV file
df_population.to_csv(csv_file_path)

# Load the population dataset from the CSV file
df_population = pd.read_csv(csv_file_path)

# Display the first few rows of the population dataset
df_population.head()

# Display a concise summary of the demographics DataFrame
df_population.info()

# Get statistical summaries of the numerical columns in the demographics DataFrame
df_population.describe()

### Key Insights from Population Data

1. **Population Distribution**:
    - The population estimates for the zip code areas vary significantly, with a mean population of approximately 20,276. This indicates a diverse range of population densities across different regions.
    - The standard deviation of 20,689 suggests a wide variation in population sizes, with some areas being densely populated while others are sparsely populated.

2. **Population Range**:
    - The minimum population estimate is 11, indicating very sparsely populated areas.
    - The maximum population estimate is 105,285, showing that some zip code areas are highly populated.
    - The interquartile range (IQR) from 1,789 to 32,942.5 highlights the central 50% of the data, suggesting that most zip code areas have populations within this range.

3. **Zip Code Coverage**:
    - The dataset covers a wide range of zip codes, from 90001 to 96161, indicating a broad geographical coverage.
    - The mean and median zip codes are close to each other, suggesting a relatively even distribution of zip codes within the dataset.

4. **Data Completeness**:
    - All columns have complete data with no missing values, ensuring the reliability of the population information for further analysis.

These insights provide a comprehensive overview of the population distribution across different zip code areas, which can be crucial for regional analysis, market segmentation, and understanding customer demographics.

## 1.2.4 IBM Services Dataset

### General Information on the IBM Services Dataset

The `telco_customer_churn_services.xlsx` dataset, abbreviated with `df_services` in our further analysis, provides detailed information about the various services subscribed to by customers of a telecommunications company. This dataset includes features such as phone service, internet service, online security, online backup, device protection, tech support, streaming TV, streaming movies, and more. Analyzing this data helps in understanding service usage patterns, customer preferences, and identifying factors that may influence customer churn.

# Load the services dataset
services_file_path = '../2_data/original_data/Telecommunications_Industry/Telco_customer_churn_services.xlsx'
csv_file_path = '../2_data/Telco_customer_churn_services.csv'
df_services = pd.read_excel(services_file_path)

# Save the services dataset to a CSV file
df_services.to_csv(csv_file_path)

# Load the services dataset from the CSV file
df_services = pd.read_csv(csv_file_path)

# Display the first few rows of the services dataset
df_services.head()

# Display a concise summary of the demographics DataFrame
df_services.info()

# Get statistical summaries of the numerical columns in the demographics DataFrame
df_services.describe()

### Key Insights from Services Data

1. **Customer Tenure**:
    - The average tenure of customers is approximately 32 months, with a standard deviation of 24.56 months. This indicates a wide range of customer loyalty, from new customers to those who have been with the company for several years.
    - The tenure ranges from 0 to 72 months, showing that the dataset includes both new and long-term customers.

2. **Monthly Charges**:
    - The average monthly charge is around $64.76, with a standard deviation of $30.09. This suggests significant variability in the services and packages subscribed to by customers.
    - Monthly charges range from $18.25 to $118.75, indicating a diverse customer base with different service levels and usage patterns.

3. **Service Subscriptions**:
    - A high proportion of customers subscribe to phone services, with many also opting for multiple lines.
    - Internet service subscriptions are common, with options including DSL, Fiber Optic, and Cable. The presence of multiple internet service types suggests varied customer preferences and availability of services.
    - Additional services such as online security, online backup, device protection, and premium tech support are also subscribed to by many customers, indicating a demand for comprehensive service packages.

4. **Streaming Services**:
    - Streaming TV, movies, and music services are popular among customers, reflecting the growing trend of using internet services for entertainment purposes.
    - The availability of unlimited data plans suggests that some customers are willing to pay extra for unrestricted data usage, likely due to high streaming and download needs.

5. **Contract and Billing Preferences**:
    - Customers have different contract types, including month-to-month, one-year, and two-year contracts. This variety allows for flexibility in customer commitments.
    - Paperless billing is a common choice, indicating a preference for digital communication and billing methods.
    - Payment methods vary, with options including bank withdrawal, credit card, and mailed check, catering to different customer preferences.

6. **Churn Analysis**:
    - The dataset includes information on customer churn, which is crucial for understanding customer retention and identifying factors that contribute to churn.
    - Analyzing the relationship between service subscriptions, tenure, monthly charges, and churn can provide valuable insights for developing strategies to reduce churn and improve customer satisfaction.

These insights highlight the diverse customer base, varied service subscriptions, and different billing preferences, which are essential for tailoring marketing strategies, improving customer retention, and enhancing overall service offerings.

## 1.2.5 IBM Status Dataset

### General Information on the IBM Status Dataset

The `telco_customer_churn_status.xlsx` dataset, abbreviated with `df_status` in our further analysis, provides detailed information about the status of customers of a telecommunications company. This dataset includes features such as tenure, contract type, payment method, and churn status. Analyzing this data helps in understanding customer retention, identifying factors that may influence customer churn, and evaluating the effectiveness of different contract types and payment methods.

# Load the status dataset
status_file_path = '../2_data/original_data/Telecommunications_Industry/Telco_customer_churn_status.xlsx'
csv_file_path = '../2_data/Telco_customer_churn_status.csv'
df_status = pd.read_excel(status_file_path)

# Save the status dataset to a CSV file
df_status.to_csv(csv_file_path)

# Load the status dataset from the CSV file
df_status = pd.read_csv(csv_file_path)

# Display the first few rows of the status dataset
df_status.head()

# Display a concise summary of the demographics DataFrame
df_status.info()

# Get statistical summaries of the numerical columns in the demographics DataFrame
df_status.describe()


### Key Insights from Status Data

1. **Customer Satisfaction**:
    - The average satisfaction score is approximately 3.24, indicating a moderate level of satisfaction among customers.
    - The satisfaction scores range from 1 (Very Unsatisfied) to 5 (Very Satisfied), with a standard deviation of 1.20, suggesting varied customer experiences.

2. **Customer Churn**:
    - About 26.54% of the customers have churned, as indicated by the mean churn value of 0.265.
    - The churn score, which predicts the likelihood of a customer leaving, has an average value of 58.51, with scores ranging from 5 to 96. This indicates a significant portion of customers are at risk of churning.

3. **Customer Lifetime Value (CLTV)**:
    - The average CLTV is 4400.30, with values ranging from 2003 to 6500. This suggests a wide range of customer values, with some customers being significantly more valuable than others.
    - The standard deviation of CLTV is 1183.06, indicating substantial variability in customer value.

4. **Churn Reasons and Categories**:
    - Only a subset of the data (1869 entries) includes specific churn reasons and categories, which can provide insights into why customers are leaving.
    - Common churn categories include Attitude, Competitor, Dissatisfaction, Other, and Price, which can help in identifying areas for improvement.

5. **Data Completeness**:
    - The dataset is complete for most columns, with the exception of 'Churn Category' and 'Churn Reason', which have non-null counts of 1869, indicating that churn reasons are only recorded for a portion of the customers.

These insights can help in understanding customer behavior, identifying at-risk customers, and improving customer retention strategies.

## 1.2.6 IBM CustomerChurn Dataset

### General Information on the IBM Customer Churn Dataset

The `CustomerChurn.xlsx` dataset, abbreviated with `df_customerc` in our further analysis, provides detailed information about the customers of a telecommunications company. This dataset includes features such as customer demographics, account information, services subscribed, and churn status. Analyzing this data helps in understanding customer retention, identifying factors that may influence customer churn, and evaluating the effectiveness of different service offerings and contract types.

# Load the status dataset
customerc_file_path = '../2_data/original_data/Telecommunications_Industry/CustomerChurn.xlsx'
csv_file_path = '../2_data/Telco_customer_churn_customerc.csv'
df_customerc = pd.read_excel(customerc_file_path)

# Save the status dataset to a CSV file
df_customerc.to_csv(csv_file_path)

# Load the status dataset from the CSV file
df_customerc = pd.read_csv(csv_file_path)

# Display the first few rows of the status dataset
df_customerc.head()

# Display a concise summary of the demographics DataFrame
df_customerc.info()

# Get statistical summaries of the numerical columns in the demographics DataFrame
df_customerc.describe()

### Key Insights from Customer Churn Data

1. **Customer Loyalty**:
    - The `LoyaltyID` column indicates a unique identifier for customer loyalty. The wide range of values suggests varying levels of customer loyalty, which can be further analyzed to understand retention patterns.

2. **Tenure**:
    - The `Tenure` column shows the number of months a customer has been with the company. The average tenure is approximately 32 months, with a standard deviation of 24.56 months. This indicates a diverse range of customer loyalty, from new customers to those who have been with the company for several years.
    - The tenure ranges from 0 to 72 months, showing that the dataset includes both new and long-term customers.

3. **Monthly Charges**:
    - The `Monthly Charges` column represents the monthly billing amount for each customer. The average monthly charge is around $64.76, with a standard deviation of $30.09. This suggests significant variability in the services and packages subscribed to by customers.
    - Monthly charges range from $18.25 to $118.75, indicating a diverse customer base with different service levels and usage patterns.

4. **Service Subscriptions**:
    - The dataset includes various service-related columns such as `Phone Service`, `Multiple Lines`, `Internet Service`, `Online Security`, `Online Backup`, `Device Protection`, `Tech Support`, `Streaming TV`, and `Streaming Movies`. These columns indicate whether customers subscribe to these services, providing insights into service usage patterns and customer preferences.

5. **Contract Types**:
    - The `Contract` column shows the type of contract customers have, including month-to-month, one-year, and two-year contracts. This variety allows for flexibility in customer commitments and can be analyzed to understand the impact of contract types on churn.

6. **Billing and Payment Methods**:
    - The `Paperless Billing` column indicates whether customers use paperless billing, and the `Payment Method` column shows the payment methods used by customers, such as electronic check, mailed check, bank transfer, and credit card. These columns provide insights into customer preferences for billing and payment, which can help tailor customer service and billing processes.

7. **Churn**:
    - The `Churn` column indicates whether customers have churned. This is the target variable for predictive modeling and is crucial for understanding customer retention. Analyzing the relationship between churn and other features can provide valuable insights for developing strategies to reduce churn and improve customer satisfaction.

These insights highlight the diverse customer base, varied service subscriptions, and different billing preferences, which are essential for tailoring marketing strategies, improving customer retention, and enhancing overall service offerings.

## 1.2.7 IBM Telco_Customer_churn Dataset

### General Information on the IBM Telco Customer Churn Dataset

The `Telco_customer_churn.xlsx` dataset, abbreviated with `df_telcocustomerc` in our further analysis, provides detailed information about the customers of a telecommunications company. This dataset includes features such as customer demographics, account information, services subscribed, and churn status. Analyzing this data helps in understanding customer retention, identifying factors that may influence customer churn, and evaluating the effectiveness of different service offerings and contract types.

# Load the services dataset
telcocustomerc_file_path = '../2_data/original_data/Telecommunications_Industry/Telco_customer_churn.xlsx'
csv_file_path = '../2_data/Telco_customer_churn_telcocustomerc.csv'
df_telcocustomerc = pd.read_excel(telcocustomerc_file_path)

# Save the services dataset to a CSV file
df_telcocustomerc.to_csv(csv_file_path)

# Load the services dataset from the CSV file
df_telcocustomerc = pd.read_csv(csv_file_path)

# Display the first few rows of the services dataset
df_telcocustomerc.head()

# Display a concise summary of the demographics DataFrame
df_telcocustomerc.info()

# Get statistical summaries of the numerical columns in the demographics DataFrame
df_telcocustomerc.describe()

### Key Insights from Telco Customer Churn Data

1. **Customer Demographics**:
  - The dataset includes demographic information such as `Gender`, `Senior Citizen`, `Partner`, and `Dependents`. This allows for analysis of how different demographic groups are distributed and how they might influence churn.

2. **Geographical Information**:
  - Columns like `Country`, `State`, `City`, `Zip Code`, `Latitude`, and `Longitude` provide detailed geographical data. This can be used to analyze regional trends and identify areas with higher churn rates.

3. **Service Details**:
  - The dataset contains information on various services subscribed by customers, including `Phone Service`, `Multiple Lines`, `Internet Service`, `Online Security`, `Online Backup`, `Device Protection`, `Tech Support`, `Streaming TV`, and `Streaming Movies`. This helps in understanding customer preferences and service usage patterns.

4. **Contract and Billing**:
  - Columns such as `Contract`, `Paperless Billing`, and `Payment Method` provide insights into the types of contracts customers have and their billing preferences. This can be crucial for analyzing the impact of contract types and billing methods on churn.

5. **Financial Information**:
  - `Monthly Charges` and `Total Charges` give an overview of the financial aspect of customer subscriptions. The wide range of charges indicates diverse service levels and customer spending patterns.

6. **Churn Information**:
  - The dataset includes several columns related to churn, such as `Churn Label`, `Churn Value`, `Churn Score`, `CLTV`, and `Churn Reason`. These columns are essential for understanding customer retention and identifying factors contributing to churn.

7. **Tenure**:
  - The `Tenure Months` column shows the duration of time customers have been with the company. The average tenure is approximately 32 months, with a wide range from 0 to 72 months, indicating a diverse customer base in terms of loyalty.

8. **Customer Lifetime Value (CLTV)**:
  - The `CLTV` column represents the customer lifetime value, which is a crucial metric for understanding the long-term value of customers. The average CLTV is around 4400, with significant variability among customers.

9. **Churn Score and Reasons**:
  - The `Churn Score` provides a numerical representation of the likelihood of a customer churning, with values ranging from 5 to 100. The `Churn Reason` column, although sparsely populated, offers qualitative insights into why customers are leaving.

10. **Data Completeness**:
   - While most columns have complete data, some columns like `Churn Reason` have a significant number of missing values. This indicates that churn reasons are only recorded for a subset of customers, which might limit the analysis of churn causes.

The Telco Customer Churn dataset provides a comprehensive view of customer demographics, geographical information, service details, financial information, and churn-related metrics. The diversity in customer profiles, service usage, and financial aspects allows for in-depth analysis of factors influencing customer churn. Understanding these insights can help in developing targeted strategies to improve customer retention and satisfaction.
