==> 3_preprocessing/7_dataset_splitting.ipynb <==
# Step 7: Split the Data

## Preprocessing Pipeline Overview

This preprocessing pipeline outlines the steps necessary to prepare the Telco Customer Churn dataset for our modeling. Each step is designed to address specific aspects of data quality, transformation, and feature creation. We will cover each step in a separate jupyter notebook file.

**Step 1: Data Loading**: Loading the datasets into the workspace, ensuring all necessary files are correctly imported for analysis. This includes the Kaggle dataset and the IBM datasets.

**Step 2: Dataset Integration**: Combining relevant datasets into a single, unified dataset that will serve as the foundation for subsequent analysis.

**Step 3: Handling Missing Values**: Identifying and addressing missing values in the dataset to ensure data integrity. This step ensures no significant gaps hinder the analysis.

**Step 4: Data Type Conversion**: Converting data columns to appropriate data types to optimize memory usage and prepare for feature engineering. Ensure consistency across all columns.

**Step 5: Data Exploration**: Perform initial exploratory data analysis (EDA) to understand the dataset's structure and characteristics, visualizing key features to gain insights into the data.

**Step 6: Feature Engineering**: Creating new features from the existing data to enhance model performance and capture additional insights. This includes transformations and derived features.

**Step 7: Dataset Splitting**: Splitting the dataset into training and testing subsets to prepare for model development and evaluation. This step ensures reproducibility and robust performance metrics.

**Step 8: Outlier Detection**: Identifying and addressing outliers in the dataset to ensure they do not negatively impact the analysis or models.

**Step 9: Clustering Customers**: Identifying the most common customer profiles via clustering.

Methods to Split the Dataset

When working with a dataset of 7034 entries, it is crucial to split the data into training and testing subsets to evaluate the models performance accurately. Here are some common methods to split the dataset:

1. **Train-Test Split**:
    - This is the most straightforward method where the dataset is divided into two parts: training and testing sets.
    - Typically, 70-80% of the data is used for training, and the remaining 20-30% is used for testing.

2. **Stratified Shuffle Split**:
    - This method ensures that the training and testing sets have the same proportion of class labels as the original dataset.
    - It is particularly useful for imbalanced datasets.

3. **K-Fold Cross-Validation**:
    - This method splits the dataset into `k` equal-sized folds. The model is trained on `k-1` folds and tested on the remaining fold.
    - This process is repeated `k` times, with each fold used exactly once as the test set.

4. **Stratified K-Fold Cross-Validation**:
    - Similar to K-Fold Cross-Validation but ensures that each fold has the same proportion of class labels.

Each of these methods has its advantages and can be chosen based on the specific requirements of the analysis and the characteristics of the dataset.


We are going to drop the 'Churn', 'Customer Status_Churned', and 'Churn Value' columns from the dataset.
The last two columns are encodings of the 'Churn' column, so all three columns are removed to avoid redundancy.


from sklearn.model_selection import StratifiedShuffleSplit
import pandas as pd

df = pd.read_csv("../2_data/telcocustomerchurn_featured.csv")
print(df.columns.tolist())
print(df.dtypes)

### Deleting Columns encoding information about Churn

We delete all columns that include "churn" in their name because these features are highly likely to introduce data leakage. 
Data leakage occurs when information from outside the training dataset is used to create the model, leading to overly optimistic performance estimates. 
In this case, columns such as 'Churn Value', 'Churn Score', 'Churn', 'Customer Status_Churned', and various 'Churn Category' and 'Churn Reason' columns 
are either explicitly tied to the target variable (churn) or derived from it. Including these columns in the model would provide it with information 
that would not be available in a real-world scenario, thus compromising the model's ability to generalize to new, unseen data.

churn_columns = [col for col in df.columns if 'Churn' in col]
print(churn_columns)

# Drop all columns containing "Churn" in their name except the "Churn" column
churn_columns_to_drop = [col for col in churn_columns if col != 'Churn']
df = df.drop(columns=churn_columns_to_drop)

# Display the remaining columns
print(df.columns.tolist())

We will drop the `unnamed_0` column, which is an additional unique identifier that was added and is not needed. Furthermore, we will drop the `Customer Status` columns, since they encode the Churn as well. Furthermore, since the 'LoyaltyID' is a numerical value encoding the Customer Loyalty (not churning), we will delete this as well.

# Define the features and target variable
X = df.drop(columns=['Customer Status_Joined', 'Customer Status_Stayed', 'LoyaltyID'])
y = df['Churn']

# StratifiedShuffleSplit 
sss = StratifiedShuffleSplit(n_splits=5, test_size=0.2, random_state=7)
print(sss)

for train_index, test_index in sss.split(X, y):
    print("train:", train_index, "test:", test_index)
    X_train, X_test = X.iloc[train_index], X.iloc[test_index]
    X_train_dashboard, X_test_dashboard = X.iloc[train_index], X.iloc[test_index]
    y_train, y_test = y.iloc[train_index], y.iloc[test_index]
    y_train_dashboard, y_test_dashboard = y.iloc[train_index], y.iloc[test_index]

    # Add 'Unnamed: 0' column to y_train_dashboard and y_test_dashboard
    y_train_dashboard = pd.DataFrame(y_train_dashboard).merge(X_train_dashboard[['Unnamed: 0']], left_index=True, right_index=True)
    y_test_dashboard = pd.DataFrame(y_test_dashboard).merge(X_test_dashboard[['Unnamed: 0']], left_index=True, right_index=True)

    # Ensure the order of rows in X_train and X_test corresponds to y_train and y_test
    

    # Sort the train and test sets to keep the order the same
    X_train = X_train.sort_index()
    X_test = X_test.sort_index()
    X_train_dashboard = X_train_dashboard.sort_index()
    X_test_dashboard = X_test_dashboard.sort_index()
    y_train = y_train.sort_index()
    y_test = y_test.sort_index()
    y_train_dashboard = y_train_dashboard.sort_index()
    y_test_dashboard = y_test_dashboard.sort_index()

print(f"X shape: {X.shape}")
print(f"X_train shape: {X_train.shape}")
print(f"X_test shape: {X_test.shape}")
print(f"y shape: {y.shape}")
print(f"y_train shape: {y_train.shape}")
print(f"y_test shape: {y_test.shape}")

# Save the train and test splits to CSV files
X_train.to_csv("../2_data/X_train.csv", index=False)
X_test.to_csv("../2_data/X_test.csv", index=False)
X_train_dashboard.to_csv("../2_data/X_train_dashboard.csv", index=False)
X_test_dashboard.to_csv("../2_data/X_test_dashboard.csv", index=False)
y_train.to_csv("../2_data/y_train.csv", index=False)
y_test.to_csv("../2_data/y_test.csv", index=False)
y_train_dashboard.to_csv("../2_data/y_train_dashboard.csv", index=False)
y_test_dashboard.to_csv("../2_data/y_test_dashboard.csv", index=False)

# Drop the specified columns from X_train and X_test
X_train = X_train.drop(columns=['Churn', 'Unnamed: 0'])
X_test = X_test.drop(columns=['Churn', 'Unnamed: 0'])
y_train = y_train.drop(columns=['Unnamed: 0'])
y_test = y_test.drop(columns=['Unnamed: 0'])

# Save the modified train and test splits to CSV files
X_train.to_csv("../2_data/X_train.csv", index=False)
X_test.to_csv("../2_data/X_test.csv", index=False)
y_train.to_csv("../2_data/y_train.csv", index=False)
y_test.to_csv("../2_data/y_test.csv", index=False)

# Load the combined dataset to get the Customer ID
combined_df = pd.read_csv("../2_data/telcocustomerchurn_combined.csv")

# Merge the Customer ID into X_train_dashboard and X_test_dashboard based on the "Unnamed: 0" column
X_train_dashboard = X_train_dashboard.merge(combined_df[['Unnamed: 0', 'Customer ID']], on='Unnamed: 0')
X_test_dashboard = X_test_dashboard.merge(combined_df[['Unnamed: 0', 'Customer ID']], on='Unnamed: 0')
y_train_dashboard = pd.DataFrame(y_train_dashboard).merge(combined_df[['Unnamed: 0', 'Customer ID']], on='Unnamed: 0')
y_test_dashboard = pd.DataFrame(y_test_dashboard).merge(combined_df[['Unnamed: 0', 'Customer ID']], on='Unnamed: 0')

X_train_dashboard = X_train_dashboard.drop(columns=['Unnamed: 0', 'Churn'])
X_test_dashboard = X_test_dashboard.drop(columns=['Unnamed: 0', 'Churn'])
y_train_dashboard = y_train_dashboard.drop(columns=['Unnamed: 0'])
y_test_dashboard = y_test_dashboard.drop(columns=['Unnamed: 0'])


# Drop any rows that have an NaN in them for X_train_dashboard, X_test_dashboard, y_train_dashboard, and y_test_dashboard
X_train_dashboard = X_train_dashboard.dropna()
X_test_dashboard = X_test_dashboard.dropna()
y_train_dashboard = y_train_dashboard.dropna()
y_test_dashboard = y_test_dashboard.dropna()



# Save the updated train and test dashboard splits to CSV files
X_train_dashboard.to_csv("../2_data/X_train_dashboard.csv", index=False)
X_test_dashboard.to_csv("../2_data/X_test_dashboard.csv", index=False)
y_train_dashboard.to_csv("../2_data/y_train_dashboard.csv", index=False)
y_test_dashboard.to_csv("../2_data/y_test_dashboard.csv", index=False)

# Print the columns of the dataframe
print(X_train.columns.tolist())


### Cross-Validation Function

By defining a cross-validation function, we can streamline the process of evaluating different models and methods, ensuring reproducibility and consistency in your results. This function can be applied to various machine learning algorithms and dataset splits, allowing you to assess their performance using the same cross-validation strategy.
We will refer to the cross validation in each method evaluation section.
